Bug 1848184 - Update lastAnalysisTime to be a long r=geckoview-reviewers,amejiamarmol,jonalmeida
Differential Revision: https://phabricator.services.mozilla.com/D186144

UltraBlame original commit: 7208a296512cf0737ea6e6c77ed46624773b1ba5

+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/GeckoSession.java
@@ -3548,7 +3548,7 @@ public class GeckoSession {
     @Nullable public final Highlight highlights;
 
     
-    public final int lastAnalysisTime;
+    public final long lastAnalysisTime;
 
     
     public final boolean deletedProductReported;
@@ -3567,7 +3567,7 @@ public class GeckoSession {
       } else {
         highlights = new Highlight(message.getBundle("highlights"));
       }
-      lastAnalysisTime = message.getInt("last_analysis_time");
+      lastAnalysisTime = message.getLong("last_analysis_time");
       deletedProductReported = message.getBoolean("deleted_product_reported");
       deletedProduct = message.getBoolean("deleted_product");
     }
